# Description

This issue fixes redirection with flash messages breaks HTTP header format by encoding msgpack serialized binary data with hex. It has little overhead, but still we get benefit of using Messagepack. Here are the benchmark results:

```go
New:
Benchmark_Redirect_Route-16                         	 6792206	       173.7 ns/op	      16 B/op	       1 allocs/op
Benchmark_Redirect_Route_WithQueries-16             	 3043641	       391.3 ns/op	      40 B/op	       2 allocs/op
Benchmark_Redirect_Route_WithFlashMessages-16       	 2439327	       481.1 ns/op	     293 B/op	       3 allocs/op
Benchmark_Redirect_parseAndClearFlashMessages-16    	 2350215	       510.5 ns/op	     192 B/op	       7 allocs/op
Benchmark_Redirect_processFlashMessages-16          	 3483320	       344.6 ns/op	     288 B/op	       2 allocs/op
Benchmark_Redirect_Messages-16                      	13537572	        87.54 ns/op	     128 B/op	       2 allocs/op
Benchmark_Redirect_OldInputs-16                     	14712171	        77.44 ns/op	      96 B/op	       2 allocs/op
Benchmark_Redirect_Message-16                       	83187558	        14.43 ns/op	       0 B/op	       0 allocs/op
Benchmark_Redirect_OldInput-16                      	153820603	         7.814 ns/op	       0 B/op	       0 allocs/op

Old:
Benchmark_Redirect_Route-16                         	 6861298	       175.9 ns/op	      16 B/op	       1 allocs/op
Benchmark_Redirect_Route_WithQueries-16             	 2901906	       408.6 ns/op	      40 B/op	       2 allocs/op
Benchmark_Redirect_Route_WithFlashMessages-16       	 3456422	       328.7 ns/op	     117 B/op	       2 allocs/op
Benchmark_Redirect_parseAndClearFlashMessages-16    	 4298755	       278.1 ns/op	      32 B/op	       6 allocs/op
Benchmark_Redirect_processFlashMessages-16          	 6022623	       198.2 ns/op	     112 B/op	       1 allocs/op
Benchmark_Redirect_Messages-16                      	13548512	        87.84 ns/op	     128 B/op	       2 allocs/op
Benchmark_Redirect_OldInputs-16                     	14798696	        78.96 ns/op	      96 B/op	       2 allocs/op
Benchmark_Redirect_Message-16                       	82884747	        14.44 ns/op	       0 B/op	       0 allocs/op
Benchmark_Redirect_OldInput-16                      	148364563	         8.104 ns/op	       0 B/op	       0 allocs/op
```

Fixes https://github.com/gofiber/fiber/issues/3437

## Type of change

- [x] Enhancement (improvement to existing features and functionality)

## Checklist

Before you submit your pull request, please make sure you meet these requirements:

- [ ] Followed the inspiration of the Express.js framework for new functionalities, making them similar in usage.
- [ ] Conducted a self-review of the code and provided comments for complex or critical parts.
- [ ] Updated the documentation in the `/docs/` directory for [Fiber's documentation](https://docs.gofiber.io/).
- [ ] Added or updated unit tests to validate the effectiveness of the changes or new features.
- [ ] Ensured that new and existing unit tests pass locally with the changes.
- [ ] Verified that any new dependencies are essential and have been agreed upon by the maintainers/community.
- [ ] Aimed for optimal performance with minimal allocations in the new code.
- [ ] Provided benchmarks for the new code to analyze and improve upon.

## Commit formatting

Please use emojis in commit messages for an easy way to identify the purpose or intention of a commit. Check out the emoji cheatsheet here: [CONTRIBUTING.md](https://github.com/gofiber/fiber/blob/master/.github/CONTRIBUTING.md#pull-requests-or-commits)