@ReneWerner87 2025-05-16
@coderabbitai can you please generate more test cases
which test the complete process coherently

i.e.
in the code a redirect with a flash message is set, then we expect as response a status code with redirect and a url and a cookie
we now send this as well as the client or browser in a 2nd request to the server with the redirect url and the cookie header from the previous response and expect that in the response by the handler our flash message from is output (this message initially came from the first handler)